chore: organize directory
